+++ CHANGELOG.md
@@ -19,6 +19,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - `actor.oldGlobalPos` returns the globalPosition from the previous frame
 - Built in actions now have a unique `id` property
+- create development builds of excalibur that bundlers can use in dev mode
+- show warning in development when Entity hasn't been added to a scene after a few seconds
 
 ### Fixed